FastBackend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pytest==8.3.5
httpx==0.28.1
pydantic==2.11.4
Pillow==11.2.1
edge-tts==7.0.2
